Table header size is now serialized as 4 bytes. Hopefully this builds and works and is happy, but I suspect I will have to fix some stuff on linux now.

+++ src/ee/storage/persistenttable.cpp
@@ -649,8 +649,8 @@ void PersistentTable::loadTuplesFrom(bool allowELT,
                                      SerializeInput &serialize_io,
                                      Pool *stringPool)
 {
-    serialize_io.readInt();
-    serialize_io.readShort();
+    serialize_io.readInt(); // table size
+    serialize_io.readInt(); // rowstart
 
     //column
     int columnCount = serialize_io.readShort(); // column num
